refactor: reorganize project structure into modular monolith
